ElasticSearch connector
- Created different "providers" for each object type; each provider knows
the relevant search-related information for its type of object. The provider
methods are absurdly simple, and if you needed to make a new one you could
probably do it by finding/replacing the name of the object. They might be a
little more complicated in the future

- Created a single "SearchData" class for providers to pass around data with
extra search-related fields, along with an "ElasticSearchWrapper" that just
knows which args/etc to use for different calls to ElasticSearch. That way
each provider only has to make 1 simple thing, and we don't have to remember
"Hmmm is it 'index' or '_index'? Do I need the timestamp for this call?" in
5+ different files

- Created a "Sync" class with all logic for syncing the db with ElasticSearch

- "run_sync.py" script actually runs the sync